// covid/src/app/models/Homework.ts
export class Homework {
    nroEmpleado: number;
    userName: string;
    nombre: string;
    apellido: string;
    unidad: string;
    donde: string;
    direccion: string;
    comuna: string;
    region: string;
    telefono: string;
    oficinaEveris: string;
    nombreCliente: string;
    direccionCliente: string;
    regionCliente: string;
    fechaInicio: Date;
    fechaFin: Date;
    sector: string;
    proyect: string;
    proyect2: string;
    comentario: string;
    cuarentena: boolean;
    inicioCuarentena: Date;
    covid19: boolean;
    fechaDiagnostico: Date;
    cliente: string;
}